// app/generated/perps/types/positionDirection.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum PositionDirection {
  Long,
  Short,
}

export type PositionDirectionArgs = PositionDirection;

export function getPositionDirectionEncoder(): FixedSizeEncoder<PositionDirectionArgs> {
  return getEnumEncoder(PositionDirection);
}

export function getPositionDirectionDecoder(): FixedSizeDecoder<PositionDirection> {
  return getEnumDecoder(PositionDirection);
}

export function getPositionDirectionCodec(): FixedSizeCodec<
  PositionDirectionArgs,
  PositionDirection
> {
  return combineCodec(
    getPositionDirectionEncoder(),
    getPositionDirectionDecoder(),
  );
}